requirements.txt
aiogram==3.15.0
aiohttp==3.11.11
feedparser==6.0.11
APScheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.2
